renamed OrderCommandHandler to TransactionHandler. changed commands to throw Exceptions instead of returning Optional Strings when errors occur. used OderStatus.valueOf instead of switch statement
